server/pyproject.toml
[tool.poetry]
name = "iqra-sbms"
version = "0.1.0"
description = "Student Billing Management System"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-pydantic = "^0.12.0"
python-dotenv = "^1.0.1"
argon2-cffi = "^23.1.0"
requests = "^2.32.3"
flask-cors = "^5.0.0"
flask-jwt-extended = "^4.6.0"
nh3 = "^0.2.18"
ruff = "^0.6.5"
flask-compress = "^1.15"
flask-sqlalchemy = "^3.1.1"
freezegun = "^1.5.1"
flask-migrate = "^4.0.7"
poetry = "^1.8.3"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
poethepoet = "^0.28.0"
pytest-order = "^1.3.0"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
indent-width = 4

[tool.ruff.format]
line-ending = "auto"
quote-style = "double"

[tool.ruff.lint]
select = ["ANN", "E", "F"]
fixable = ["ALL"]
ignore = ["ANN101"]

[tool.poe.tasks]
run-dev = "poetry run python run.py development"
run-dev-flask = "poetry run python -m flask run --host=0.0.0.0 --port=5001 --debug"
run-prod = "gunicorn -w 4 wsgi:app"
test = "pytest ./src/tests --import-mode=append --order-scope=session -v -s"
clean = "poetry run python scripts/cleanup.py"
lint = "ruff format"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"